ggseg3d: Plot brain segmentations with plotly

Plotting results from analyses done on data derived from brain segmentations is a common need, but may be quite laborious. Results from such analyses are usually easier to interpret if the plot can mimmick the shape and position in the brain it represents.

ggseg3d plots and returns a plotly mesh3d object.

ggseg3d(
  .data = NULL,
  atlas = "dk_3d",
  surface = "LCBC",
  hemisphere = c("right", "subcort"),
  label = "region",
  text = NULL,
  colour = "colour",
  palette = NULL,
  na.colour = "darkgrey",
  na.alpha = 1,
  show.legend = TRUE,
  options.legend = NULL,
  ...
)

Arguments

.data
A data.frame to use for plot aesthetics. Must include a column called "region" corresponding to regions.

atlas
Either a string with the name of a 3d atlas to use.

surface
String. Which surface to plot. Either "pial","white", or "inflated"[default]

hemisphere
String. Hemisphere to plot. Either "left" or "right"[default], can also be "subcort".

label
String. Quoted name of column in atlas/data that should be used to name traces

text
String. Quoted name of column in atlas/data that should be added as extra information in the hover text.

colour
String. Quoted name of column from which colour should be supplied

palette
String. Vector of colour names or HEX colours. Can also be a named numeric vector, with colours as names, and breakpoint for that colour as the value

na.colour
String. Either name, hex of RGB for colour of NA in colour.

na.alpha
Numeric. A number between 0 and 1 to control transparency of NA-regions.

show.legend
Logical. Toggle legend if colour is numeric.

options.legend
list of layout changes to colour bar

...
additional arguments to add_trace

Available surfaces:

`inflated:` Fully inflated surface

`semi-inflated:` Semi-inflated surface

`white:` white matter surface

Value

a plotly tri-surface mesh plot

Details

This package contains data from various brain parcellations, with convenient functions to inspect the results directly on a brain-plot wither with plotly in 3d meshes.

See also

Author

Maintainer: Athanasia Mo Mowinckel a.m.mowinckel@psykologi.uio.no (ORCID)

Authors:

Didac Vidal-Piñeiro d.v.pineiro@psykologi.uio.no (ORCID)

Athanasia Mowinckel and Didac Piñeiro

Examples

ggseg3d()
ggseg3d(hemisphere = "left")
ggseg3d(surface = "inflated")
ggseg3d(show.legend = FALSE)